# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Recommendation API",
    "description": "Stubbed recommendation API for PSB recommender platform.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/recommendations": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Recommendations",
        "operationId": "get_recommendations_api_v1_recommendations_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "User Id"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Channel",
              "description": "Origin channel"
            },
            "description": "Origin channel"
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Number of recommendations",
              "default": 5,
              "title": "Top K"
            },
            "description": "Number of recommendations"
          },
          {
            "name": "x-correlation-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/feedback": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Post Feedback",
        "operationId": "post_feedback_api_v1_recommendations_feedback_post",
        "parameters": [
          {
            "name": "x-correlation-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Channel": {
        "type": "string",
        "enum": [
          "mobile",
          "web",
          "crm"
        ],
        "title": "Channel"
      },
      "EventType": {
        "type": "string",
        "enum": [
          "click",
          "dismiss",
          "purchase"
        ],
        "title": "EventType"
      },
      "FeedbackRequest": {
        "properties": {
          "rec_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Rec Id"
          },
          "user_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "User Id"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "context": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "rec_id",
          "user_id",
          "event_type"
        ],
        "title": "FeedbackRequest"
      },
      "FeedbackResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "correlation_id"
        ],
        "title": "FeedbackResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RecommendationItem": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "reasons": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasons"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "product_name",
          "score"
        ],
        "title": "RecommendationItem"
      },
      "RecommendationResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationItem"
            },
            "type": "array",
            "title": "Recommendations"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "generated_at",
          "correlation_id",
          "recommendations"
        ],
        "title": "RecommendationResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "recommendations",
      "description": "Recommendation operations"
    },
    {
      "name": "feedback",
      "description": "Feedback ingestion"
    }
  ]
}